Move examples to src/Export
