+++ src/FINDOLOGIC/Export/Examples/XmlExample.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace FINDOLOGIC\Examples;
+namespace FINDOLOGIC\Export\Examples;
 
 use FINDOLOGIC\Export\Data\Attribute;
 use FINDOLOGIC\Export\Data\Image;
@@ -12,17 +12,15 @@ use FINDOLOGIC\Export\Data\Usergroup;
 use FINDOLOGIC\Export\Exporter;
 
 /**
- * This example class builds a CSV export inspired by the FINDOLOGIC documentation, which can be found at
- * https://docs.findologic.com/doku.php?id=export_patterns:csv .
+ * This example class builds a xml export based on the example of the FINDOLOGIC documentation, which can be found
+ * under the following link https://docs.findologic.com/doku.php?id=export_patterns:xml#example_xml
  * @codeCoverageIgnore
  */
-class CsvExample
+class XmlExample
 {
     public function createExport(): string
     {
-        $exporter = Exporter::create(Exporter::TYPE_CSV, 20, [
-            'sale', 'novelty', 'logo', 'availability', 'old_price', 'Basic_rate_price'
-        ]);
+        $exporter = Exporter::create(Exporter::TYPE_XML);
 
         $item = $exporter->createItem('01120c948ad41a2284ad9f0402fbc7d');
 
@@ -75,16 +73,22 @@ class CsvExample
     private function addBonuses(Item $item): void
     {
         $item->addBonus(3);
+        $item->addBonus(5, 'LNrLF7BRVJ0toQ==');
     }
 
     private function addDateAddeds(Item $item): void
     {
         $item->addDateAdded(new \DateTime());
+        $item->addDateAdded(new \DateTime(), 'LNrLF7BRVJ0toQ==');
     }
 
     private function addDescriptions(Item $item): void
     {
         $item->addDescription('With this sneaker you will walk in style. It\'s available in green and blue.');
+        $item->addDescription(
+            'With this men\'s sneaker you will walk in style. It\'s comes in various sizes and colors.',
+            'LNrLF7BRVJ0toQ=='
+        );
     }
 
     private function addOrdernumbers(Item $item): void
@@ -93,6 +97,9 @@ class CsvExample
             '' => [
                 '277KTL',
                 '4987123846879'
+            ],
+            'LNrLF7BRVJ0toQ==' => [
+                '377KTL'
             ]
         ];
 
@@ -107,7 +114,12 @@ class CsvExample
     {
         $imagesData = [
             '' => [
-                'https://www.store.com/images/277KTL.png' => Image::TYPE_DEFAULT
+                'https://www.store.com/images/277KTL.png' => Image::TYPE_DEFAULT,
+                'https://www.store.com/images/thumbnails/277KTL.png' => Image::TYPE_THUMBNAIL
+            ],
+            'LNrLF7BRVJ0toQ==' => [
+                'https://www.store.com/images/277KTLmen.png' => Image::TYPE_DEFAULT,
+                'https://www.store.com/images/thumbnails/277KTLmen.png' => Image::TYPE_THUMBNAIL
             ]
         ];
 
@@ -124,6 +136,9 @@ class CsvExample
             '' => [
                 '277KTL',
                 '4987123846879'
+            ],
+            'LNrLF7BRVJ0toQ==' => [
+                '377KTL'
             ]
         ];
 
@@ -137,36 +152,41 @@ class CsvExample
     private function addNames(Item $item): void
     {
         $item->addName('Adidas Sneaker');
+        $item->addName('Adidas Men\'s Sneaker', 'LNrLF7BRVJ0toQ==');
     }
 
     private function addPrices(Item $item): void
     {
         $item->addPrice(44.8);
-        $item->setInsteadPrice(50);
-        $item->setMaxPrice(47);
-        $item->setTaxRate(20);
+        $item->addPrice(45.9, 'LNrLF7BRVJ0toQ==');
     }
 
     private function addProperties(Item $item): void
     {
         $propertiesData = [
             'sale' => [
-                '' => 1
+                '' => 1,
+                'LNrLF7BRVJ0toQ==' => 0
             ],
             'novelty' => [
-                '' => 0
+                '' => 0,
+                'LNrLF7BRVJ0toQ==' => 0
             ],
             'logo' => [
-                '' => 'http://www.shop.de/brand.png'
+                '' => 'http://www.shop.de/brand.png',
+                'LNrLF7BRVJ0toQ==' => 'http://www.shop.de/brand.png'
             ],
             'availability' => [
-                '' => '<span style="color: green;">4 days</span>'
+                '' => '<span style="color: green;">4 days</span>',
+                'LNrLF7BRVJ0toQ==' => '<span style="color: green;">3 days</span>'
             ],
             'old_price' => [
-                '' => 99.9
+                '' => 99.9,
+                'LNrLF7BRVJ0toQ==' => 99.9
             ],
             'Basic_rate_price' => [
-                '' => 99.9
+                '' => 99.9,
+                'LNrLF7BRVJ0toQ==' => 89.9
             ]
         ];
 
@@ -179,21 +199,25 @@ class CsvExample
     private function addSalesFrequencies(Item $item): void
     {
         $item->addSalesFrequency(5);
+        $item->addSalesFrequency(10, 'LNrLF7BRVJ0toQ==');
     }
 
     private function addSorts(Item $item): void
     {
         $item->addSort(5);
+        $item->addSort(7, 'LNrLF7BRVJ0toQ==');
     }
 
     private function addSummaries(Item $item): void
     {
         $item->addSummary('A cool and fashionable sneaker');
+        $item->addSummary('A cool and fashionable sneaker for men', 'LNrLF7BRVJ0toQ==');
     }
 
     private function addUrls(Item $item): void
     {
         $item->addUrl('https://www.store.com/sneakers/adidas.html');
+        $item->addUrl('https://www.store.com/sneakers/mens/adidas.html', 'LNrLF7BRVJ0toQ==');
     }
 
     private function addUsergroups(Item $item): void
